Get Companies from Health Canada Drug Product Database

Retrieves information on all pharmaceutical companies listed in the Health Canada Drug Product Database (DPD) using the RESTful API endpoint /drug/company.

This includes details such as the company code, company name, address, city, province, postal code, and country. Each record corresponds to a company associated with one or more approved or discontinued drug products.

Usage

get_hc_companies()

Source

Health Canada Drug Product Database (DPD) API: https://health-products.canada.ca/api/documentation/dpd-documentation-en.html

Value

A tibble with the following columns:

company_code: Unique identifier for the company

company_name: Official registered name of the company

company_type: Type of company (e.g., "DIN OWNER", "Manufacturer")

city_name: City where the company is located

province_name: Province or territory (if applicable)

country_name: Country name

postal_code: Postal code or ZIP code

street_name: Street address

Details

This function sends a GET request to the Health Canada Drug Product Database API. It supports caching via the memoise package to avoid redundant calls, and includes a small rate limit delay between successive API requests.

The columns post_office_box and suite_number are automatically removed as they generally contain incomplete or irrelevant information.

If the API request fails or returns an error status code, the function returns NULL with an informative message.

Note

Requires an active internet connection.

See also

Examples

# \donttest{
  # This function requires an internet connection and downloads data from Health Canada
  get_hc_companies()
#> # A tibble: 5,219 × 8
#>    city_name     company_code company_name company_type country_name postal_code
#>    <chr>                <int> <chr>        <chr>        <chr>        <chr>      
#>  1 Englewood Cl…        13667 SKINSCIENCE… DIN OWNER    United Stat… 07632      
#>  2 Minneapolis           3692 AVEDA CORPO… DIN OWNER    United Stat… 55449      
#>  3 Boucherville         11498 WAMPOLE BRA… DIN OWNER    Canada       J4B 5H2    
#>  4 Hacienta Hei…         9791 GRAND HEALT… DIN OWNER    United Stat… 91745      
#>  5 Scarborough           5755 CLAYMORE PR… DIN OWNER    Canada       M1V 5C8    
#>  6 Charlotte             5395 WE CARE DIS… DIN OWNER    United Stat… 28208      
#>  7 Toronto               3644 ARAMIS, DIS… DIN OWNER    Canada       M3J 2W3    
#>  8 West Columbia         9039 NEPHRON PHA… DIN OWNER    United Stat… 29172      
#>  9 Halifax              18106 THE ORB FAC… DIN OWNER    Canada       B3P 1L3    
#> 10 Vaughan               9107 LIGI INC.    DIN OWNER    Canada       L4K 4P1    
#> # ℹ 5,209 more rows
#> # ℹ 2 more variables: province_name <chr>, street_name <chr>
# }